automatically backed up by learn

+++ lib/bachelor.rb
@@ -57,5 +57,5 @@ def get_average_age_for_season(data, season)
         end
     end
   end
-  return ages.inject{ |sum, el| sum + el }.to_f / ages.size
+  return ages.inject{ |sum, el| sum + el }.to_i / ages.size
 end
